Moved Picking data sources to separate folder
